Interim commit. Finished factoring partitioning stuff out of point cloud stuff.
