+++ core/exchanger.c
@@ -1213,317 +1213,3 @@ void exchanger_transfer_metadata(exchanger_t* ex,
   exchanger_finish_metadata_transfer(ex, token);
 }
 
-struct migrator_t
-{
-  exchanger_t* ex;
-};
-
-static void migrator_free(void* ctx)
-{
-  migrator_t* m = ctx;
-  m->ex = NULL;
-}
-
-migrator_t* migrator_new(MPI_Comm comm)
-{
-  migrator_t* m = polymec_gc_malloc(sizeof(migrator_t), migrator_free);
-  m->ex = exchanger_new(comm);
-  return m;
-}
-
-migrator_t* migrator_clone(migrator_t* m)
-{
-  migrator_t* clone = polymec_gc_malloc(sizeof(migrator_t), migrator_free);
-  clone->ex = exchanger_clone(m->ex);
-  return clone;
-}
-
-int migrator_num_sends(migrator_t* m)
-{
-  return exchanger_num_sends(m->ex);
-}
-
-bool migrator_next_send(migrator_t* m, int* pos, int* remote_process, int** indices, int* num_indices)
-{
-  return exchanger_next_send(m->ex, pos, remote_process, indices, num_indices);
-}
-
-int migrator_num_receives(migrator_t* m)
-{
-  return exchanger_num_receives(m->ex);
-}
-
-bool migrator_next_receive(migrator_t* m, int* pos, int* remote_process, int** indices, int* num_indices)
-{
-  return exchanger_next_receive(m->ex, pos, remote_process, indices, num_indices);
-}
-
-void migrator_verify(migrator_t* m, void (*handler)(const char* format, ...))
-{
-  exchanger_verify(m->ex, handler);
-}
-
-migrator_t* migrator_from_global_partition(MPI_Comm comm, 
-                                           int64_t* global_partition,
-                                           size_t num_global_vertices)
-{
-  START_FUNCTION_TIMER();
-  int nprocs, rank;
-  MPI_Comm_size(comm, &nprocs);
-  MPI_Comm_rank(comm, &rank);
-
-  exchanger_t* ex = exchanger_new(comm);
-  
-  if (rank == 0)
-  {
-    // Get the number of vertices we're going to send to each other process.
-    size_t num_vertices_to_send[nprocs];
-    memset(num_vertices_to_send, 0, sizeof(size_t) * nprocs);
-    for (int v = 0; v < (int)num_global_vertices; ++v)
-      num_vertices_to_send[global_partition[v]]++;
-
-    // Send this number.
-    size_t n;
-    MPI_Scatter(num_vertices_to_send, 1, MPI_SIZE_T, &n, 1, MPI_SIZE_T, 0, comm);
-
-    // Now send the vertices to each process and register these sends
-    // with the distributor.
-    for (int p = 1; p < nprocs; ++p)
-    {
-      ASSERT(num_vertices_to_send[p] > 0);
-      int vertices[num_vertices_to_send[p]], k = 0;
-      for (int i = 0; i < (int)num_global_vertices; ++i)
-      {
-        if (global_partition[i] == p)
-          vertices[k++] = i;
-      }
-      exchanger_set_send(ex, p, vertices, (int)num_vertices_to_send[p], true);
-    }
-
-    // Figure out the local vertices for rank 0.
-    size_t num_local_vertices = num_vertices_to_send[0];
-    int local_vertices[num_local_vertices], k = 0;
-    for (int i = 0; i < (int)num_global_vertices; ++i)
-    {
-      if (global_partition[i] == 0)
-        local_vertices[k++] = i;
-    }
-  }
-  else
-  {
-    // Get the number of vertices we will receive from rank 0.
-    size_t num_local_vertices;
-    MPI_Scatter(NULL, 1, MPI_SIZE_T, &num_local_vertices, 1, MPI_SIZE_T, 0, comm);
-
-    // Now get the vertices.
-    int local_vertices[num_local_vertices];
-    for (int i = 0; i < (int)num_local_vertices; ++i)
-      local_vertices[i] = i;
-
-    // Now register all the vertices we're receiving with the distributor.
-    ASSERT(num_local_vertices > 0);
-    exchanger_set_receive(ex, 0, local_vertices, (int)num_local_vertices, true);
-  }
-
-  migrator_t* m = polymec_gc_malloc(sizeof(migrator_t), migrator_free);
-  m->ex = ex;
-  STOP_FUNCTION_TIMER();
-  return m;
-}
-
-migrator_t* migrator_from_local_partition(MPI_Comm comm,
-                                          int64_t* local_partition,
-                                          size_t num_vertices)
-{
-  START_FUNCTION_TIMER();
-  exchanger_t* ex = exchanger_new(comm);
-  
-  // Tally up the number of vertices we're going to send to every other process, 
-  // including those that are staying on our own.
-  int nprocs, rank;
-  MPI_Comm_size(comm, &nprocs);
-  MPI_Comm_rank(comm, &rank);
-  int num_vertices_to_send[nprocs];
-  memset(num_vertices_to_send, 0, sizeof(int) * nprocs);
-  for (int v = 0; v < (int)num_vertices; ++v)
-    num_vertices_to_send[local_partition[v]]++;
-
-  // Get the number of vertices we're going to receive from every other process.
-  int num_vertices_to_receive[nprocs];
-  MPI_Alltoall(num_vertices_to_send, 1, MPI_INT, 
-               num_vertices_to_receive, 1, MPI_INT, comm);
-
-  // Set up send and receive vertices.
-  int num_local_vert = num_vertices_to_receive[rank];
-  int recv_offset = 0;
-  for (int p = 0; p < nprocs; ++p)
-  {
-    if (p != rank)
-    {
-      if ((num_vertices_to_receive[p] > 0) && (p != rank))
-      {
-        if (p != rank)
-        {
-          int num_vert = num_vertices_to_receive[p];
-          int receive_vertices[num_vertices_to_receive[p]];
-          for (int i = 0; i < num_vert; ++i, ++recv_offset)
-            receive_vertices[i] = num_local_vert + recv_offset;
-          exchanger_set_receive(ex, p, receive_vertices, num_vertices_to_receive[p], true);
-        }
-      }
-      if (num_vertices_to_send[p] > 0)
-      {
-        int send_vertices[num_vertices_to_send[p]], s = 0;
-        for (int v = 0; v < num_vertices; ++v)
-        {
-          if (local_partition[v] == p)
-            send_vertices[s++] = v;
-        }
-        if (p != rank)
-          exchanger_set_send(ex, p, send_vertices, num_vertices_to_send[p], true);
-      }
-    }
-  }
-
-  migrator_t* m = polymec_gc_malloc(sizeof(migrator_t), migrator_free);
-  m->ex = ex;
-  STOP_FUNCTION_TIMER();
-  return m;
-}
-
-void migrator_transfer(migrator_t* m, void* data, int* count, int stride, int tag, MPI_Datatype type)
-{
-  START_FUNCTION_TIMER();
-  int token = migrator_start_transfer(m, data, count, stride, tag, type);
-  migrator_finish_transfer(m, token);
-  STOP_FUNCTION_TIMER();
-}
-
-int migrator_start_transfer(migrator_t* m, void* data, int* count, int stride, int tag, MPI_Datatype type)
-{
-#if POLYMEC_HAVE_MPI
-  // This is the same as an exchange, with an additional datum (the length 
-  // of data) stored.
-  int token = exchanger_start_exchange(m->ex, data, stride, tag, type);
-  m->ex->transfer_counts[token] = count; 
-  return token;
-#else
-  return 0;
-#endif
-}
-
-void migrator_finish_transfer(migrator_t* m, int token)
-{
-#if POLYMEC_HAVE_MPI
-  START_FUNCTION_TIMER();
-  ASSERT(token >= 0);
-  ASSERT(token < m->ex->num_pending_msgs);
-  ASSERT(m->ex->transfer_counts[token] != NULL); // We can't finish an exchange as a transfer!
-
-  // Retrieve the message and its size (in data elements) for the given token.
-  mpi_message_t* msg = m->ex->pending_msgs[token];
-  int* count = m->ex->transfer_counts[token];
-
-  // Wait for the transfer to finish.
-  int err = exchanger_waitall(m->ex, msg);
-  if (err != MPI_SUCCESS) return;
-
-  // Copy the received data into the original array.
-  void* orig_buffer = m->ex->orig_buffers[token];
-  mpi_message_unpack(msg, orig_buffer, m->ex->receive_offset, m->ex->receive_map);
-
-  // Count up the elements we actually received.
-  int num_received = 0, pos = 0, proc;
-  exchanger_channel_t* c;
-  while (exchanger_map_next(m->ex->receive_map, &pos, &proc, &c))
-    num_received += c->num_indices;
-
-  // We need to kill all of the sent elements that aren't overwritten by received elements.
-  // So first make a list of all the received indices.
-  int num_sent = 0;
-  int_unordered_set_t* overwritten_sends = int_unordered_set_new();
-  pos = 0;
-  while (exchanger_map_next(m->ex->send_map, &pos, &proc, &c))
-  {
-    for (int i = 0; i < c->num_indices; ++i)
-      int_unordered_set_insert(overwritten_sends, c->indices[i]);
-    num_sent += c->num_indices;
-  }
-
-  // Update the count.
-  *count += (num_received - num_sent);
-  ASSERT(*count >= 0);
-  if (num_sent - num_received > 0)
-  {
-    log_debug("migrator_transfer: deleting %d elements locally.", 
-              num_sent - num_received);
-  }
-  else if (num_sent - num_received < 0)
-  {
-    log_debug("migrator_transfer: adding %d elements locally.", 
-              num_received - num_sent);
-  }
-
-  // Record the "width" in bytes of the type we're transferring.
-  size_t width = mpi_size(msg->type);
-
-  // Remove the send elements that aren't overwritten by received ones.
-  // We recreate the array in buffer, skipping over the elements we're removing.
-  char* array = orig_buffer;
-  int stride = msg->stride, offset = 0;
-  char* buffer = polymec_malloc(width * stride * (*count));
-  pos = 0;
-  while (exchanger_map_next(m->ex->send_map, &pos, &proc, &c))
-  {
-    for (int i = 0; i < c->num_indices; ++i)
-    {
-      int index = c->indices[i];
-      if (!int_unordered_set_contains(overwritten_sends, index))
-      {
-        memcpy(&buffer[stride*width*offset], &array[stride*width*index], stride*width);
-        ++offset;
-      }
-    }
-  }
-  memcpy(array, buffer, width*stride*offset);
-  polymec_free(buffer);
-  int_unordered_set_free(overwritten_sends);
-
-  // Pull the message out of our list of pending messages and delete it.
-  m->ex->pending_msgs[token] = NULL;
-  m->ex->orig_buffers[token] = NULL;
-  m->ex->transfer_counts[token] = NULL;
-  mpi_message_free(msg);
-  STOP_FUNCTION_TIMER();
-#endif
-}
-
-void migrator_fprintf(migrator_t* m, FILE* stream)
-{
-  fprintf(stream, "Migrator: ");
-  exchanger_fprintf(m->ex, stream);
-}
-
-static size_t m_size(void* obj)
-{
-  return ex_size(obj);
-}
-
-static void* m_read(byte_array_t* bytes, size_t* offset)
-{
-  migrator_t* m = polymec_gc_malloc(sizeof(migrator_t), migrator_free);
-  m->ex = ex_read(bytes, offset);
-  return m;
-}
-
-static void m_write(void* obj, byte_array_t* bytes, size_t* offset)
-{
-  migrator_t* m = obj;
-  ex_write(m->ex, bytes, offset);
-}
-
-serializer_t* migrator_serializer()
-{
-  return serializer_new("migrator", m_size, m_read, m_write, NULL);
-}
-
