+++ core/exchanger.h
@@ -207,75 +207,4 @@ int exchanger_start_metadata_transfer(exchanger_t* ex,
 void exchanger_finish_metadata_transfer(exchanger_t* ex,
                                         int token);
 
-//------------------------------------------------------------------------
-//                            Data migration
-//------------------------------------------------------------------------
-// When a computational domain is repartitioned or reorganized , data often 
-// needs to be migrated from one process to another. Migrator objects help 
-// with this task. They are very similer to exchangers but move data instead
-// of exchanging it per se.
-//------------------------------------------------------------------------
-
-// Migrator class -- related to the exchanger class, but used differently.
-typedef struct migrator_t migrator_t;
-
-// Creates a migrator that does nothing, for the given MPI communicator.
-migrator_t* migrator_new(MPI_Comm comm);
-
-// This function creates a migrator that can distribute data from the 
-// root process (0) according to the global partition vector. The partition 
-// vector is NULL on all nonzero ranks and defined on rank 0.
-migrator_t* migrator_from_global_partition(MPI_Comm comm, 
-                                           int64_t* global_partition,
-                                           size_t num_global_vertices);
-
-// This function creates a migrator that can migrate data from all 
-// processes to others according to their respective local partition vectors.
-migrator_t* migrator_from_local_partition(MPI_Comm comm,
-                                          int64_t* local_partition,
-                                          size_t num_local_vertices);
-
-// Creates a complete copy of the given migrator.
-migrator_t* migrator_clone(migrator_t* m);
-
-// Returns the number of processes to which this migrator sends data.
-int migrator_num_sends(migrator_t* m);
-
-// Allows the traversal of the set of the migrator's send indices for remote 
-// processes.
-bool migrator_next_send(migrator_t* m, int* pos, int* remote_process, int** indices, int* num_indices);
-
-// Returns the number of processes from which this migrator receives data.
-int migrator_num_receives(migrator_t* m);
-
-// Allows the traversal of the set of the migrator's receive indices for 
-// remote processes.
-bool migrator_next_receive(migrator_t* m, int* pos, int* remote_process, int** indices, int* num_indices);
-
-// Verifies the consistency of the migrator, raising an error in the 
-// case of inconsistency, using the given error handler. Involves parallel 
-// communication, so make sure it is called by all processes on the 
-// communicator for the migrator.
-void migrator_verify(migrator_t* m, void (*handler)(const char* format, ...));
-
-// Transfers data between processes, creating new received elements 
-// and deleting old sent elements where needed. The array data initially 
-// contains a number of elements compatible with the exchanger, while the 
-// final number of elements in data (the initial count minus those sent 
-// elements, which are jettisoned) is stored in count.
-void migrator_transfer(migrator_t* m, void* data, int* count, int stride, int tag, MPI_Datatype type);
-
-// Begins the asynchronous transfer of data between processes, returning
-// a unique token.
-int migrator_start_transfer(migrator_t* m, void* data, int* count, int stride, int tag, MPI_Datatype type);
-
-// Concludes the asynchronous transfer of data between processes.
-void migrator_finish_transfer(migrator_t* m, int token);
-
-// This writes a string representation of the migrator to the given file stream.
-void migrator_fprintf(migrator_t* m, FILE* stream);
-
-// This creates a serializer object that can read and write migrators to byte streams.
-serializer_t* migrator_serializer(void);
-
 #endif
