+++ core/tests/CMakeLists.txt
@@ -47,7 +47,6 @@ add_polymec_core_test(test_cxx_program test_cxx_program.cpp)
 add_mpi_polymec_core_test(test_parallel_timers test_timers.c 1 2 3 4)
 add_mpi_polymec_core_test(test_parallel_kd_tree test_kd_tree.c 1 2 3 4)
 add_mpi_polymec_core_test(test_exchanger test_exchanger.c 1 2 3 4)
-add_mpi_polymec_core_test(test_migrator test_migrator.c 1 2 3 4)
 add_mpi_polymec_core_test(test_partition_point_cloud test_partition_point_cloud.c ../../geometry/create_point_lattice.c ../../geometry/cubic_lattice.c 1 2 3 4)
 add_mpi_polymec_core_test(test_parallel_sort test_parallel_sort.c 1 2 3 4)
 add_mpi_polymec_core_test(test_repartition_point_cloud test_repartition_point_cloud.c 1 2 3 4)
