+++ core/tests/test_partition_point_cloud.c
@@ -30,8 +30,7 @@ static void test_partition_linear_cloud(void** state, int N)
   }
 
   // Partition it.
-  migrator_t* m = partition_point_cloud(&cloud, comm, NULL, 0.05);
-  m = NULL;
+  assert_true(partition_point_cloud(&cloud, comm, NULL, 0.05, NULL, 0));
 
   // Now check data. Points should all fall on dx tick marks.
   for (int i = 0; i < cloud->num_points; ++i)
@@ -65,8 +64,7 @@ static void test_partition_planar_cloud(void** state, int nx, int ny)
   }
 
   // Partition it.
-  migrator_t* m = partition_point_cloud(&cloud, comm, NULL, 0.05);
-  m = NULL;
+  assert_true(partition_point_cloud(&cloud, comm, NULL, 0.05, NULL, 0));
 
   // Clean up.
   point_cloud_free(cloud);
@@ -88,8 +86,7 @@ static void test_partition_cubic_cloud(void** state, int nx, int ny, int nz)
   }
 
   // Partition it.
-  migrator_t* m = partition_point_cloud(&cloud, comm, NULL, 0.05);
-  m = NULL;
+  assert_true(partition_point_cloud(&cloud, comm, NULL, 0.05, NULL, 0));
 
   // Clean up.
   point_cloud_free(cloud);
